src/requirements.txt
aws-cdk-lib==2.222.0   # replace with the latest stable version
constructs==10.4.3     # compatible range for CDK v2
flask==3.1.2           # used to define web services